add more comments on the app to help the user
I implemented one of the feedbacks we received during the lab which was to add some text to our app to give more information about the graphs, the data,... So I added a text that explains where the data comes from. I also added an introduction that explains why the app is useful and how it can help the user. I added a text next to the map to explain what it represents. I also added 1/2 small sentences aboves each graph to explain what they are representing.